fix route ids
