+++ packages/insomnia/src/ui/components/modals/request-settings-modal.tsx
@@ -11,8 +11,8 @@ import type { SocketIORequest } from '../../../models/socket-io-request';
 import { isWebSocketRequest, type WebSocketRequest } from '../../../models/websocket-request';
 import { invariant } from '../../../utils/invariant';
 import { useRequestPatcher } from '../../hooks/use-request';
-import type { ListWorkspacesLoaderData } from '../../routes/$organizationId.project';
-import { revalidateWorkspaceActiveRequest } from '../../routes/$organizationId.project.$projectId.workspace';
+import type { ListWorkspacesLoaderData } from '../../routes/$organizationId.project.$projectId';
+import { revalidateWorkspaceActiveRequest } from '../../routes/$organizationId.project.$projectId.workspace.$workspaceId';
 import { Modal, type ModalHandle, type ModalProps } from '../base/modal';
 import { ModalBody } from '../base/modal-body';
 import { ModalHeader } from '../base/modal-header';
