+++ packages/insomnia/src/ui/components/modals/workspace-certificates-modal.tsx
@@ -19,7 +19,7 @@ import {
 import { useFetcher, useParams, useRouteLoaderData } from 'react-router';
 
 import type { ClientCertificate } from '../../../models/client-certificate';
-import type { WorkspaceLoaderData } from '../../routes/$organizationId.project.$projectId.workspace';
+import type { WorkspaceLoaderData } from '../../routes/$organizationId.project.$projectId.workspace.$workspaceId';
 import { Icon } from '../icon';
 import { PasswordViewer } from '../viewers/password-viewer';
