+++ packages/insomnia/src/ui/context/app/insomnia-event-stream-context.tsx
@@ -7,8 +7,8 @@ import type { Organization } from '../../../models/organization';
 import { VCSInstance } from '../../../sync/vcs/insomnia-sync';
 import { insomniaFetch } from '../../../ui/insomniaFetch';
 import { avatarImageCache } from '../../hooks/image-cache';
-import type { ProjectIdLoaderData } from '../../routes/$organizationId.project';
-import type { WorkspaceLoaderData } from '../../routes/$organizationId.project.$projectId.workspace';
+import type { ProjectIdLoaderData } from '../../routes/$organizationId.project.$projectId';
+import type { WorkspaceLoaderData } from '../../routes/$organizationId.project.$projectId.workspace.$workspaceId';
 import { useRootLoaderData } from '../../routes/root';
 
 const InsomniaEventStreamContext = createContext<{
