+++ packages/insomnia/src/ui/hooks/use-runner-request-list.tsx
@@ -5,7 +5,7 @@ import { isRequest, type Request } from '../../models/request';
 import { isRequestGroup } from '../../models/request-group';
 import { invariant } from '../../utils/invariant';
 import { useRunnerContext } from '../context/app/runner-context';
-import type { Child, WorkspaceLoaderData } from '../routes/$organizationId.project.$projectId.workspace';
+import type { Child, WorkspaceLoaderData } from '../routes/$organizationId.project.$projectId.workspace.$workspaceId';
 import type { RequestRow } from '../routes/$organizationId.project.$projectId.workspace.$workspaceId.debug.runner';
 
 export const useRunnerRequestList = (organizationId: string, targetFolderId: string, runnerId: string) => {
