+++ packages/insomnia/src/ui/hooks/use-vcs-version.ts
@@ -4,8 +4,8 @@ import { useParams } from 'react-router';
 
 import { type ChangeBufferEvent, database } from '../../common/database';
 import type { BaseModel } from '../../models';
-import type { ProjectLoaderData } from '../routes/$organizationId.project';
-import type { WorkspaceLoaderData } from '../routes/$organizationId.project.$projectId.workspace';
+import type { ProjectLoaderData } from '../routes/$organizationId.project.$projectId';
+import type { WorkspaceLoaderData } from '../routes/$organizationId.project.$projectId.workspace.$workspaceId';
 // We use this hook to determine if the active request has been updated from the system (not the user typing)
 // For example, by pulling a new version from the remote, switching branches, etc.
 export function useActiveRequestSyncVCSVersion() {
