+++ packages/insomnia/src/ui/routes/organization.sync-organizations-and-projects.tsx
@@ -7,7 +7,7 @@ import type { Project } from '../../models/project';
 import { invariant } from '../../utils/invariant';
 import { AsyncTask } from '../../utils/router';
 import { migrateProjectsUnderOrganization, syncOrganizations } from '../organization-utils';
-import { syncProjects } from './$organizationId.project';
+import { syncProjects } from './$organizationId.project.$projectId';
 
 interface SyncOrgsAndProjectsActionRequest {
   organizationId: string;
